+++ Cargo.toml
@@ -1,35 +1,5 @@
-[package]
-name = "pubnub"
-version = "0.1.0"
-description = "PubNub Async lib based on Tokio"
-repository = "https://github.com/pubnub/rust"
-keywords = ["pubnub", "publish-subscribe", "subscribe"]
-authors = ["Stephen Blum <stephen@pubnub.com>", "Jay Oster <jay@pubnub.com>"]
-readme = "readme.md"
-edition = "2018"
-
-[dependencies]
-error-iter = "=0.2.0"
-futures-executor-preview = "=0.3.0-alpha.19"
-futures-util-preview = { version = "=0.3.0-alpha.19", features = ["async-await", "select-macro"] }
-json = "=0.12.0"
-log = "=0.4.8"
-percent-encoding = "=2.1.0"
-thiserror = "=1.0.6"
-http = "=0.1.21"
-async-trait = "=0.1.21"
-
-tokio = { version = "=0.2.0-alpha.6", optional = true }
-hyper = { version = "=0.13.0-alpha.4", optional = true }
-hyper-tls = { version = "=0.4.0-alpha.4", optional = true }
-
-[dev-dependencies]
-byteorder = "1.3"
-env_logger = "0.7"
-getrandom = "0.1"
-randomize = "3.0"
-
-[features]
-default = ["transport_hyper"]
-transport_hyper = ["runtime_tokio", "hyper", "hyper-tls"]
-runtime_tokio = ["tokio"]
+[workspace]
+members = [
+  "pubnub-core",
+  "pubnub-hyper"
+]
