+++ pubnub-hyper/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "pubnub-hyper"
+version = "0.1.0"
+description = "PubNub async crate based on Tokio and Hyper"
+repository = "https://github.com/pubnub/rust"
+keywords = ["pubnub", "publish-subscribe", "subscribe"]
+authors = ["Stephen Blum <stephen@pubnub.com>", "Jay Oster <jay@pubnub.com>"]
+edition = "2018"
+readme = "../README.md"
+
+[dependencies]
+pubnub-core = { path = "../pubnub-core" }
+
+tokio = "=0.2.0-alpha.6"
+hyper = "=0.13.0-alpha.4"
+hyper-tls = "=0.4.0-alpha.4"
+
+async-trait = "=0.1.21"
+error-iter = "=0.2.0"
+thiserror = "=1.0.6"
+
+[dev-dependencies]
+byteorder = "1.3"
+env_logger = "0.7"
+getrandom = "0.1"
+randomize = "3.0"
+futures-executor-preview = "=0.3.0-alpha.19"
+futures-util-preview = { version = "=0.3.0-alpha.19", features = ["async-await", "select-macro"] }
+json = "=0.12.0"
+log = "=0.4.8"
+percent-encoding = "=2.1.0"
+http = "=0.1.21"
+tokio = "=0.2.0-alpha.6"
