+++ pubnub-hyper/src/adapters/mod.rs
