+++ pubnub-hyper/src/adapters/transport/hyper/error.rs
@@ -1,3 +1,4 @@
+use crate::core::json;
 use error_iter::ErrorIter;
 use thiserror::Error;
