+++ pubnub-hyper/src/adapters/transport/hyper/mod.rs
@@ -1,5 +1,6 @@
-use crate::message::{Message, Timetoken, Type};
-use crate::transport::Transport as Trait;
+use crate::core::json;
+use crate::core::Transport as Trait;
+use crate::core::{Message, Timetoken, Type};
 
 use async_trait::async_trait;
