+++ pubnub-hyper/src/adapters/transport/mod.rs
@@ -1,0 +1,1 @@
+pub mod hyper;
