--- src/adapters/runtime/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "runtime_tokio")]
-pub mod tokio;
