--- src/adapters/transport/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "transport_hyper")]
-pub mod hyper;
